Refactor: Naming, tests, typing
