+++ pontos/github/api/api.py
@@ -32,17 +32,17 @@ from pontos.github.api.labels import GitHubRESTLabelsMixin
 from pontos.github.api.organizations import GitHubRESTOrganizationsMixin
 from pontos.github.api.pull_requests import GitHubRESTPullRequestsMixin
 from pontos.github.api.release import GitHubRESTReleaseMixin
-from pontos.github.api.workflows import GitHubAPIWorkflowsMixin
+from pontos.github.api.workflows import GitHubRESTWorkflowsMixin
 
 
 class GitHubRESTApi(
-    GitHubRESTPullRequestsMixin,
-    GitHubRESTReleaseMixin,
     GitHubRESTArtifactsMixin,
     GitHubRESTBranchMixin,
     GitHubRESTLabelsMixin,
-    GitHubAPIWorkflowsMixin,
     GitHubRESTOrganizationsMixin,
+    GitHubRESTPullRequestsMixin,
+    GitHubRESTReleaseMixin,
+    GitHubRESTWorkflowsMixin,
 ):
     """GitHubRESTApi Mixin"""
